feat: 로또 티켓 번호 정렬
