+++ src/main/java/lotto/README.md
@@ -24,3 +24,4 @@
 
 - [x] 보너스 볼을 입력 받는다.
 - [x] 5개 일치 + 보너스 볼 일치의 경우 2등 당첨
+- [x] 로또 티켓 번호 정렬
